Only make the query to get all tables once
On larger datasets, the query was timing out
or taking a very long time, and making that
query every 100 records was taking forever to get
through the script. This will still fail with too
many records, as the _table_metadata view does not
do a true limit, it has to get all the data first
and then give back the first n rows. So the query
takes ages regardless of the limit.